requirements.txt
flask
flask-cors
gunicorn
PyPDF2
docx2txt
requests
pandas
python-dotenv





